// libs/chart/src/lib/__generated__/Candles.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { Interval } from "@vegaprotocol/graphql";

// ====================================================
// GraphQL query operation: Candles
// ====================================================

export interface Candles_market_tradableInstrument_instrument {
  __typename: 'Instrument';
  /**
   * Uniquely identify an instrument across all instruments available on Vega (string)
   */
  id: string;
  /**
   * Full and fairly descriptive name for the instrument
   */
  name: string;
  /**
   * A short non necessarily unique code used to easily describe the instrument (e.g: FX:BTCUSD/DEC18) (string)
   */
  code: string;
}

export interface Candles_market_tradableInstrument {
  __typename: 'TradableInstrument';
  /**
   * An instance of or reference to a fully specified instrument.
   */
  instrument: Candles_market_tradableInstrument_instrument;
}

export interface Candles_market_candles {
  __typename: 'Candle';
  /**
   * RFC3339Nano formatted date and time for the candle
   */
  datetime: string;
  /**
   * High price (uint64)
   */
  high: string;
  /**
   * Low price (uint64)
   */
  low: string;
  /**
   * Open price (uint64)
   */
  open: string;
  /**
   * Close price (uint64)
   */
  close: string;
  /**
   * Volume price (uint64)
   */
  volume: string;
}

export interface Candles_market {
  __typename: 'Market';
  /**
   * Market ID
   */
  id: string;
  /**
   * decimalPlaces indicates the number of decimal places that an integer must be shifted by in order to get a correct
   * number denominated in the currency of the Market. (uint64)
   *
   * Examples:
   * Currency     Balance  decimalPlaces  Real Balance
   * GBP              100              0       GBP 100
   * GBP              100              2       GBP   1.00
   * GBP              100              4       GBP   0.01
   * GBP                1              4       GBP   0.0001   (  0.01p  )
   *
   * GBX (pence)      100              0       GBP   1.00     (100p     )
   * GBX (pence)      100              2       GBP   0.01     (  1p     )
   * GBX (pence)      100              4       GBP   0.0001   (  0.01p  )
   * GBX (pence)        1              4       GBP   0.000001 (  0.0001p)
   */
  decimalPlaces: number;
  /**
   * An instance of or reference to a tradable instrument.
   */
  tradableInstrument: Candles_market_tradableInstrument;
  /**
   * Candles on a market, for the 'last' n candles, at 'interval' seconds as specified by params
   */
  candles: (Candles_market_candles | null)[] | null;
}

export interface Candles {
  /**
   * An instrument that is trading on the VEGA network
   */
  market: Candles_market | null;
}

export interface CandlesVariables {
  marketId: string;
  interval: Interval;
  since: string;
}